Extract maintenance module from store + break down enqueue + complete.
This is the final commit in a series of commits in which we've been
breaking down the massive `store` module into smaller modules. What was
over 13K lines of code is now spread across many modules with < 1K each.
`store.rs` itself is now well under 1K lines.

In this commit, `maintenance.rs` is extracted, which includes the
compaction and backup functionality, and the `enqueue` and `complete`
modules have their `mod.rs` broken into smaller modules by functionality
area.

No further work is on the immediate agenda for breaking down the store.
